@@ -1,10 +1,7 @@
 class Group < ApplicationRecord
-  validates :name, :presence => true
+  validates :name, presence: :true
 
-  def is_member?
-    if self.user_id
-      user = User.find(self.user_id)
-      !!user.first_name
-    end
+  def members
+    User.find(member_ids)
   end
 end
